refactor | sprint2 | FRB-180 | 기존 서비스와 레포 서비스 분리(Equip,Sensor,Zone,Worker,AbnormalLog) 리팩토링 | 정민석
