+++ src/main/java/com/factoreal/backend/domain/abnormalLog/application/AbnormalLogRepoService.java
@@ -1,0 +1,21 @@
+package com.factoreal.backend.domain.abnormalLog.application;
+
+import com.factoreal.backend.domain.abnormalLog.dao.AbnLogRepository;
+import com.factoreal.backend.domain.abnormalLog.entity.AbnormalLog;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class AbnormalLogRepoService {
+
+    private AbnLogRepository abnLogRepository;
+
+    public Page<AbnormalLog> findByZone_ZoneIdOrderByDetectedAtDesc(String zoneId, Pageable pageable) {
+        return abnLogRepository.findByZone_ZoneIdOrderByDetectedAtDesc(zoneId, pageable);
+    }
+}
