+++ src/main/java/com/factoreal/backend/domain/sensor/application/SensorService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import com.factoreal.backend.domain.equip.application.EquipRepoService;
 import com.factoreal.backend.domain.equip.application.EquipService;
 import com.factoreal.backend.domain.sensor.dto.request.SensorCreateRequest;
 import com.factoreal.backend.domain.sensor.dto.response.SensorInfoResponse;
 import com.factoreal.backend.domain.sensor.dto.request.SensorUpdateRequest;
+import com.factoreal.backend.domain.zone.application.ZoneRepoService;
 import com.factoreal.backend.domain.zone.application.ZoneService;
 import com.factoreal.backend.messaging.kafka.strategy.enums.SensorType;
 import org.springframework.http.HttpStatus;
@@ -27,9 +29,11 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 @RequiredArgsConstructor
 public class SensorService {
-    private final SensorRepository sensorRepository;
-    private final ZoneService zoneService;
-    private final EquipService equipService;
+    private final SensorRepoService sensorRepoService;
+//    private final ZoneService zoneService;
+//    private final EquipService equipService;
+    private final ZoneRepoService zoneRepoService;
+    private final EquipRepoService equipRepoService;
 
     /**
      * 센서 정보 저장, 하드코딩으로 센서를 저장하고자 할때 사용
@@ -45,24 +49,14 @@ public class SensorService {
         sens.setZone(zone);
         sens.setEquip(equip);
         sens.setIsZone(dto.getIsZone());
-        return save(sens);
+        return sensorRepoService.save(sens);
     }
 
     /**
-     * 모든 센서 리스트 조회
+     * 모든 센서 리스트 조회 서비스
      */
     public List<SensorInfoResponse> getAllSensors() {
-        return sensorRepository.findAll().stream()
-                .map(s -> new SensorInfoResponse(
-                        s.getSensorId(),
-                        s.getSensorType().toString(),
-                        s.getZone().getZoneId(),
-                        s.getEquip().getEquipId(),
-                        s.getSensorThres(),
-                        s.getAllowVal(),
-                        s.getIsZone()
-                ))
-                .collect(Collectors.toList());
+        return sensorRepoService.findAll();
     }
 
     /**
@@ -70,55 +64,55 @@ public class SensorService {
      */
     @Transactional
     public void updateSensor(String sensorId, SensorUpdateRequest dto) {
-        Sensor sensor = getSensorById(sensorId);
+        Sensor sensor = sensorRepoService.getSensorById(sensorId);
         sensor.setSensorThres(dto.getSensorThres());
         sensor.setAllowVal(dto.getAllowVal());
-        save(sensor);
+        sensorRepoService.save(sensor);
     }
 
+//    /**
+//     * 센서 정보 저장을 위한 레포 접근 메서드
+//     */
+//    @Transactional
+//    protected Sensor save(Sensor sensor) {
+//        return sensorRepository.save(sensor);
+//    }
+    
     /**
-     * 센서 정보 저장을 위한 레포 접근 메서드
-     */
-    @Transactional
-    protected Sensor save(Sensor sensor) {
-        return sensorRepository.save(sensor);
-    }
-
-    /**
-     * 공간 Id로 공간 조회하는 레포 접근 메서드
+     * 공간 Id로 공간 조회하는 메서드
      */
     private Zone getZoneById(String zoneId) {
-        return zoneService.findById(zoneId);
+        return zoneRepoService.findById(zoneId);
     }
 
     /**
-     * 설비 Id로 설비를 조회하는 레포 접근 메서드
+     * 설비 Id로 설비를 조회하는 메서드
      */
     private Equip getEquipById(String eqiuipId) {
-        return equipService.findById(eqiuipId);
+        return equipRepoService.findById(eqiuipId);
     }
 
-    /**
-     * 센서 Id로 센서를 조회하는 레포 접근 메서드
-     */
-    public Sensor getSensorById(String sensorId) {
-        return sensorRepository.findById(sensorId)
-                .orElseThrow(() -> new ResponseStatusException(
-                        HttpStatus.NOT_FOUND, "존재하지 않는 센서 ID: " + sensorId));
-    }
+//    /**
+//     * 센서 Id로 센서를 조회하는 레포 접근 메서드
+//     */
+//    public Sensor getSensorById(String sensorId) {
+//        return sensorRepository.findById(sensorId)
+//                .orElseThrow(() -> new ResponseStatusException(
+//                        HttpStatus.NOT_FOUND, "존재하지 않는 센서 ID: " + sensorId));
+//    }
 
-    /**
-     * 공간에 존재하는 센서 리스트를 조회하는 레포 접근 메서드
-     */
-    public List<Sensor> findByZone(Zone zone) {
-        return sensorRepository.findByZone(zone);
-    }
-
-    public Optional<Sensor> findById(String sensorId) {
-        return sensorRepository.findById(sensorId);
-    }
-
-    public List<Sensor> findByZone_ZoneId(String zoneId) {
-        return sensorRepository.findByZone_ZoneId(zoneId);
-    }
+//    /**
+//     * 공간에 존재하는 센서 리스트를 조회하는 레포 접근 메서드
+//     */
+//    public List<Sensor> findByZone(Zone zone) {
+//        return sensorRepository.findByZone(zone);
+//    }
+//
+//    public Optional<Sensor> findById(String sensorId) {
+//        return sensorRepository.findById(sensorId);
+//    }
+//
+//    public List<Sensor> findByZone_ZoneId(String zoneId) {
+//        return sensorRepository.findByZone_ZoneId(zoneId);
+//    }
 }
