+++ src/main/java/com/factoreal/backend/domain/worker/application/WorkerRepoService.java
@@ -1,0 +1,20 @@
+package com.factoreal.backend.domain.worker.application;
+
+import com.factoreal.backend.domain.worker.dao.WorkerRepository;
+import com.factoreal.backend.domain.worker.entity.Worker;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class WorkerRepoService {
+    private final WorkerRepository workerRepository;
+
+    public Worker findById(String workerId) {
+        return workerRepository.findById(workerId)
+                .orElseThrow(() -> new IllegalArgumentException("작업자를 찾을 수 없습니다: " + workerId));
+    }
+
+}
