+++ src/main/java/com/factoreal/backend/messaging/grafana/GrafanaZoneService.java
@@ -1,5 +1,6 @@
 package com.factoreal.backend.messaging.grafana;
 
+import com.factoreal.backend.domain.sensor.application.SensorRepoService;
 import com.factoreal.backend.domain.sensor.application.SensorService;
 import com.factoreal.backend.domain.sensor.entity.Sensor;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -14,9 +15,10 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class GrafanaZoneService {
-    private final SensorService sensorService;
+//    private final SensorService sensorService;
     private final DashboardFactory dashboardFactory;
     private final GrafanaClient grafanaClient;
+    private final SensorRepoService sensorRepoService;
 
     @Value("${grafana.url}")
     private String grafanaUrl;
@@ -36,7 +38,7 @@ public class GrafanaZoneService {
      */
     public List<GrafanaSensorResponseDto> createDashboardUrls(String zoneId) throws JsonProcessingException {
         // 1) 센서 목록 조회
-        List<Sensor> sensors = sensorService.findByZone_ZoneId(zoneId);
+        List<Sensor> sensors = sensorRepoService.findByZone_ZoneId(zoneId);
         if (sensors.isEmpty()) {
             throw new IllegalStateException("No sensors for zone: " + zoneId);
         }
